requirements.txt
# GUI Framework
customtkinter==5.2.2

# Database
mysql-connector-python==9.1.0
PyMySQL==1.1.0

# API and HTTP Requests
requests==2.32.3

# Data Visualization
matplotlib==3.9.2
pillow==11.0.0

# Data Processing
pandas==2.2.3
numpy==2.1.3

# Configuration and Environment
python-dotenv==1.0.0

# Date and Time
python-dateutil==2.8.2

# Password Hashing
bcrypt==4.1.2

# Utilities
python-slugify==8.0.1

# Logging (built-in but optional enhancement)
colorlog==6.8.0

# Development and Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0